docstring and names
