+++ docs/apidocs/ibmq-provider.rst
@@ -12,3 +12,4 @@ Qiskit IBM Quantum Provider API Reference
    ibmq_visualization
    ibmq_jupyter
    ibmq_utils
+   ibmq_random
