+++ docs/apidocs/ibmq_random.rst
@@ -1,0 +1,6 @@
+.. _qiskit-providers-ibmq-random:
+
+.. automodule:: qiskit.providers.ibmq.random
+   :no-members:
+   :no-inherited-members:
+   :no-special-members:
