+++ qiskit/providers/ibmq/credentials/credentials.py
@@ -79,7 +79,7 @@ class Credentials:
 
         # Initialize additional service URLs.
         services = services or {}
-        self.cqc_url = services.get('extractorsService', None)
+        self.extractor_url = services.get('extractorsService', None)
 
     def is_ibmq(self) -> bool:
         """Return whether the credentials represent an IBM Quantum Experience account."""
