+++ qiskit/providers/ibmq/random/__init__.py
@@ -21,6 +21,17 @@ Random Number Services (:mod:`qiskit.providers.ibmq.random`)
 
 Modules related to IBM Quantum Experience random number generator services.
 
+.. caution::
+
+  This package is currently provided in beta form and heavy modifications to
+  both functionality and API are likely to occur.
+
+The only service currently provided is the Cambridge Quantum Computing (CQC)
+extractor. To use this service, you need to first generate raw random bits
+and a set of parameters for the extractor. See
+`qiskit_rng <https://github.com/qiskit-community/qiskit_rng>`_ for more
+details.
+
 Classes
 ==========================
 .. autosummary::
